requirements.txt
fastapi==0.75.2
uvicorn[standard]==0.17.6
beautifulsoup4==4.11.1
httpx==0.22.0
lxml==4.8.0
feedparser==6.0.8
slowapi==0.1.5
tweepy==4.8.0
starlette==0.17.1
python-dotenv==0.20.0